+++ benches/genomic.rs
@@ -5,7 +5,7 @@ use symagen::random_data;
 use abd_clam::{
     cakes::{RnnAlgorithm, CAKES},
     cluster::PartitionCriteria,
-    dataset::VecVec,
+    dataset::VecDataset,
     COMMON_METRICS_STR,
 };
 
@@ -36,7 +36,7 @@ fn genomic(c: &mut Criterion) {
 
         println!("Building cakes for {metric_name} ...");
         let data_name = format!("{metric_name}-{cardinality}");
-        let dataset = VecVec::new(data.clone(), metric, data_name, true);
+        let dataset = VecDataset::new(data.clone(), metric, data_name, true);
         let criteria = PartitionCriteria::new(true).with_min_cardinality(1);
         let cakes = CAKES::new(dataset, Some(seed), criteria);
