+++ examples/genomic.rs
@@ -6,7 +6,7 @@ use symagen::random_data;
 use abd_clam::{
     cakes::{RnnAlgorithm, CAKES},
     cluster::PartitionCriteria,
-    dataset::VecVec,
+    dataset::VecDataset,
 };
 
 fn main() {
@@ -18,7 +18,7 @@ fn main() {
     let data = data.iter().map(|v| v.as_str()).collect();
 
     let metric = levenshtein::<u16>;
-    let data = VecVec::new(data, metric, "genomic".to_string(), true);
+    let data = VecDataset::new(data, metric, "genomic".to_string(), true);
 
     let criteria = PartitionCriteria::new(true).with_min_cardinality(1);
     let model = CAKES::new(data, Some(42), criteria);
