+++ src/core/cluster/criteria.rs
@@ -4,20 +4,36 @@ use distances::Number;
 
 use super::Cluster;
 
-// Note (OWM): This leaks cluster if we allow it to be public. Getting this to make sense is a TODO
+/// A criterion used to decide when to partition a `Cluster`.
+///
+/// # Type Parameters
+///
+/// - `T`: The type of the instances in the `Cluster`.
+/// - `U`: The type of the distance values between instances.
 pub trait PartitionCriterion<T: Send + Sync + Copy, U: Number>: std::fmt::Debug + Send + Sync {
-    // TODO (Najib): figure out how not to lean Cluster here
+    /// Check whether a `Cluster` meets the criterion for partitioning.
+    // TODO: Figure out how to have this not leak `Cluster` or make `Cluster` public.
     fn check(&self, c: &Cluster<T, U>) -> bool;
 }
 
+/// A collection of criteria used to decide when to partition a `Cluster`.
 #[allow(clippy::module_name_repetitions)]
 #[derive(Debug)]
 pub struct PartitionCriteria<T: Send + Sync + Copy, U: Number> {
+    /// The criteria used to decide when to partition a `Cluster`.
     criteria: Vec<Box<dyn PartitionCriterion<T, U>>>,
+    /// Whether all criteria must be met for a `Cluster` to be partitioned or if any one criterion
+    /// is sufficient.
     check_all: bool,
 }
 
 impl<T: Send + Sync + Copy, U: Number> PartitionCriteria<T, U> {
+    /// Create a new `PartitionCriteria` instance.
+    ///
+    /// # Arguments
+    ///
+    /// * `check_all`: if `true`, all criteria must be met for a `Cluster` to be partitioned, if
+    /// `false`, any one criterion is sufficient.
     #[must_use]
     pub fn new(check_all: bool) -> Self {
         Self {
@@ -26,24 +42,40 @@ impl<T: Send + Sync + Copy, U: Number> PartitionCriteria<T, U> {
         }
     }
 
+    /// Add the `MaxDepth` criterion to the collection of criteria.
+    ///
+    /// # Arguments
+    ///
+    /// * `threshold`: the maximum depth of a `Cluster` beyond which it may not be partitioned.
     #[must_use]
     pub fn with_max_depth(mut self, threshold: usize) -> Self {
         self.criteria.push(Box::new(MaxDepth(threshold)));
         self
     }
 
+    /// Add the `MinCardinality` criterion to the collection of criteria.
+    ///
+    /// # Arguments
+    ///
+    /// * `threshold`: the minimum cardinality of a `Cluster` below which it may not be partitioned.
     #[must_use]
     pub fn with_min_cardinality(mut self, threshold: usize) -> Self {
         self.criteria.push(Box::new(MinCardinality(threshold)));
         self
     }
 
+    /// Add a custom criterion to the collection of criteria.
+    ///
+    /// # Arguments
+    ///
+    /// * `c`: the custom criterion to add.
     #[allow(dead_code)]
     pub(crate) fn with_custom(mut self, c: Box<dyn PartitionCriterion<T, U>>) -> Self {
         self.criteria.push(c);
         self
     }
 
+    /// Check whether a `Cluster` meets the criteria for partitioning.
     pub(crate) fn check(&self, cluster: &Cluster<T, U>) -> bool {
         !cluster.is_singleton()
             && if self.check_all {
@@ -54,8 +86,9 @@ impl<T: Send + Sync + Copy, U: Number> PartitionCriteria<T, U> {
     }
 }
 
+/// The maximum depth of a `Cluster` beyond which it may not be partitioned.
 #[derive(Debug, Clone)]
-struct MaxDepth(usize);
+pub struct MaxDepth(usize);
 
 impl<T: Send + Sync + Copy, U: Number> PartitionCriterion<T, U> for MaxDepth {
     fn check(&self, c: &Cluster<T, U>) -> bool {
@@ -63,8 +96,9 @@ impl<T: Send + Sync + Copy, U: Number> PartitionCriterion<T, U> for MaxDepth {
     }
 }
 
+/// The minimum cardinality of a `Cluster` below which it may not be partitioned.
 #[derive(Debug, Clone)]
-struct MinCardinality(usize);
+pub struct MinCardinality(usize);
 
 impl<T: Send + Sync + Copy, U: Number> PartitionCriterion<T, U> for MinCardinality {
     fn check(&self, c: &Cluster<T, U>) -> bool {
