+++ src/core/cluster/tree.rs
@@ -1,4 +1,4 @@
-use std::marker::PhantomData;
+//! A `Tree` represents a hierarchy of "similar" instances from a metric-`Space`.
 
 use distances::Number;
 
@@ -6,16 +6,26 @@ use crate::dataset::Dataset;
 
 use super::{Cluster, PartitionCriteria};
 
-/// A `Tree` represents a hierarchy of "similar" instances from a
-/// metric-`Space`.
+/// A `Tree` represents a hierarchy of "similar" instances from a metric-`Space`.
 ///
 /// Typically one will chain calls to `new`, `build`, and finally
 /// `partition` to construct a fully realized `Tree`.
+///
+/// # Type Parameters
+///
+/// - `T`: The type of the instances in the `Tree`.
+/// - `U`: The type of the distance values between instances.
+/// - `D`: The type of the `Dataset` from which the `Tree` is built.
 #[derive(Debug)]
 pub struct Tree<T: Send + Sync + Copy, U: Number, D: Dataset<T, U>> {
+    /// The dataset from which the tree is built.
     data: D,
+    /// The root `Cluster` of the tree.
     root: Cluster<T, U>,
-    _t: PhantomData<T>,
+    /// The depth of the tree.
+    depth: usize,
+    /// The type of the instances in the tree.
+    center: T,
 }
 
 impl<T: Send + Sync + Copy, U: Number, D: Dataset<T, U>> Tree<T, U, D> {
@@ -25,11 +35,20 @@ impl<T: Send + Sync + Copy, U: Number, D: Dataset<T, U>> Tree<T, U, D> {
     /// # Arguments
     /// dataset: The dataset from which the tree will be built
     pub fn new(data: D, seed: Option<u64>) -> Self {
+        let root = Cluster::new_root(&data, data.indices(), seed);
+        let depth = root.max_leaf_depth();
+        let center = root.center;
         Self {
-            root: Cluster::new_root(&data, data.indices(), seed),
             data,
-            _t: PhantomData::default(),
+            root,
+            depth,
+            center,
         }
+    }
+
+    /// Returns a reference to dataset associated with the tree
+    pub const fn data(&self) -> &D {
+        &self.data
     }
 
     /// # Returns
@@ -38,9 +57,16 @@ impl<T: Send + Sync + Copy, U: Number, D: Dataset<T, U>> Tree<T, U, D> {
         &self.root
     }
 
-    /// Returns a reference to dataset associated with the tree
-    pub const fn data(&self) -> &D {
-        &self.data
+    /// # Returns
+    /// A reference to the root `Cluster` of the tree
+    pub(crate) const fn depth(&self) -> usize {
+        self.depth
+    }
+
+    /// # Returns
+    /// A reference to the root `Cluster` of the tree
+    pub(crate) const fn center(&self) -> T {
+        self.center
     }
 
     /// # Returns
