+++ src/core/mod.rs
@@ -1,3 +1,5 @@
+//! Core modules for the crate.
+
 pub mod cluster;
 pub mod dataset;
 // pub mod graph;
