+++ src/lib.rs
@@ -1,6 +1,6 @@
 #![deny(clippy::correctness)]
 #![warn(
-    // missing_docs,
+    missing_docs,
     clippy::all,
     clippy::suspicious,
     clippy::style,
@@ -8,7 +8,7 @@
     clippy::perf,
     clippy::pedantic,
     clippy::nursery,
-    // clippy::missing_docs_in_private_items,
+    clippy::missing_docs_in_private_items,
     // clippy::unwrap_used,
     clippy::expect_used,
     clippy::panic,
@@ -24,8 +24,10 @@ pub(crate) mod utils;
 
 pub use crate::core::{cluster, dataset};
 
+/// The current version of the crate.
 pub const VERSION: &str = "0.17.0";
 
+/// Common distance functions and their names for slices of `f32`.
 #[allow(clippy::type_complexity)]
 pub const COMMON_METRICS_F32: &[(&str, fn(&[f32], &[f32]) -> f32)] = &[
     ("euclidean", distances::vectors::euclidean),
@@ -34,6 +36,7 @@ pub const COMMON_METRICS_F32: &[(&str, fn(&[f32], &[f32]) -> f32)] = &[
     ("cosine", distances::vectors::cosine),
 ];
 
+/// Common distance functions and their names for slices of `f64`.
 #[allow(clippy::type_complexity)]
 pub const COMMON_METRICS_F64: &[(&str, fn(&[f64], &[f64]) -> f64)] = &[
     ("euclidean", distances::vectors::euclidean),
@@ -42,6 +45,7 @@ pub const COMMON_METRICS_F64: &[(&str, fn(&[f64], &[f64]) -> f64)] = &[
     ("cosine", distances::vectors::cosine),
 ];
 
+/// Common distance functions and their names for `&str`.
 #[allow(clippy::type_complexity)]
 pub const COMMON_METRICS_STR: &[(&str, fn(&str, &str) -> u32)] = &[
     ("hamming", distances::strings::hamming),
