+++ src/utils/mod.rs
@@ -1,1 +1,3 @@
+//! Utility functions for the crate.
+
 pub mod helpers;
